+++ picasso/gui/render.py
@@ -13,7 +13,6 @@ import copy
 import time
 import os.path
 import importlib, pkgutil
-from glob import glob
 from math import ceil
 from functools import partial
 
@@ -48,12 +47,6 @@ if IMSWRITER:
     from .. ext.bitplane import numpy_to_imaris
     from PyImarisWriter.ImarisWriterCtypes import *
     from PyImarisWriter import PyImarisWriter as PW
-
-try:
-    from hdbscan import HDBSCAN
-    HDBSCAN_IMPORTED = True
-except ModuleNotFoundError:
-    HDBSCAN_IMPORTED = False
 
 if sys.platform == "darwin": # plots do not work on mac os
     matplotlib.use('agg')
@@ -208,10 +201,7 @@ def check_picks(f):
             QtWidgets.QMessageBox.information(
                 args[0],
                 "Pick Error",
-                (
-                    "No localizations picked."
-                    " Please pick at least twice first."
-                ),
+                "Please pick at least twice.",
             )
         else:
             return f(args[0])
@@ -561,10 +551,10 @@ class DatasetDialog(QtWidgets.QDialog):
         self.auto_display = QtWidgets.QCheckBox("Automatic display update")
         self.auto_display.setChecked(True)
         self.auto_colors = QtWidgets.QCheckBox("Automatic coloring")
-        self.layout.addWidget(self.legend, 0, 0)
-        self.layout.addWidget(self.auto_display, 1, 0)
-        self.layout.addWidget(self.wbackground, 2, 0)
-        self.layout.addWidget(self.auto_colors, 3, 0)
+        self.layout.addWidget(self.legend, 0, 0, 1, 2)
+        self.layout.addWidget(self.auto_display, 1, 0, 1, 2)
+        self.layout.addWidget(self.wbackground, 2, 0, 1, 2)
+        self.layout.addWidget(self.auto_colors, 3, 0, 1, 2)
         self.layout.addWidget(QtWidgets.QLabel("Files"), 4, 0)
         self.layout.addWidget(QtWidgets.QLabel("Change title"), 4, 1)
         self.layout.addWidget(QtWidgets.QLabel("Color"), 4, 2)
@@ -889,7 +879,7 @@ class DatasetDialog(QtWidgets.QDialog):
             ".hdf5", "_colors.txt"
         )
         path, ext = QtWidgets.QFileDialog.getSaveFileName(
-            self, "Save colors to", out_path, filter="*txt"
+            self, "Save colors to", out_path, filter="*.txt"
         )
         if path:
             with open(path, "w") as file:
@@ -1417,11 +1407,10 @@ class ClsDlg3D(QtWidgets.QDialog):
             + "."
         )
 
-        print("Mode 1")
         locs = all_picked_locs[current]
         locs = stack_arrays(locs, asrecarray=True, usemask=False)
 
-        est = KMeans(n_clusters=n_clusters)
+        est = KMeans(n_clusters=n_clusters, n_init='auto')
 
         scaled_locs = lib.append_to_rec(
             locs, locs["x"] * pixelsize, "x_scaled"
@@ -1441,7 +1430,9 @@ class ClsDlg3D(QtWidgets.QDialog):
         counts = list(Counter(labels).items())
         # l_locs = lib.append_to_rec(l_locs,labels,'cluster')
 
-        ax1.scatter(locs["x"], locs["y"], locs["z"], c=labels.astype(np.float), s=2)
+        ax1.scatter(
+            locs["x"], locs["y"], locs["z"], c=labels.astype(float), s=2
+        )
 
         ax1.set_xlabel("X")
         ax1.set_ylabel("Y")
@@ -1615,11 +1606,10 @@ class ClsDlg2D(QtWidgets.QDialog):
             + "."
         )
 
-        print("Mode 1")
         locs = all_picked_locs[current]
         locs = stack_arrays(locs, asrecarray=True, usemask=False)
 
-        est = KMeans(n_clusters=n_clusters)
+        est = KMeans(n_clusters=n_clusters, n_init='auto')
 
         scaled_locs = lib.append_to_rec(locs, locs["x"], "x_scaled")
         scaled_locs = lib.append_to_rec(scaled_locs, locs["y"], "y_scaled")
@@ -1634,7 +1624,7 @@ class ClsDlg2D(QtWidgets.QDialog):
         counts = list(Counter(labels).items())
         # l_locs = lib.append_to_rec(l_locs,labels,'cluster')
 
-        ax1.scatter(locs["x"], locs["y"], c=labels.astype(np.float), s=2)
+        ax1.scatter(locs["x"], locs["y"], c=labels.astype(float), s=2)
 
         ax1.set_xlabel("X")
         ax1.set_ylabel("Y")
@@ -1790,7 +1780,7 @@ class DbscanDialog(QtWidgets.QDialog):
         grid.addWidget(self.radius, 0, 1)
         grid.addWidget(QtWidgets.QLabel("Min. samples:"), 1, 0)
         self.density = QtWidgets.QSpinBox()
-        self.density.setRange(1, 1e6)
+        self.density.setRange(1, int(1e6))
         self.density.setValue(4)
         grid.addWidget(self.density, 1, 1)
         vbox.addLayout(grid)
@@ -1859,12 +1849,12 @@ class HdbscanDialog(QtWidgets.QDialog):
         grid = QtWidgets.QGridLayout()
         grid.addWidget(QtWidgets.QLabel("Min. cluster size:"), 0, 0)
         self.min_cluster = QtWidgets.QSpinBox()
-        self.min_cluster.setRange(1, 1e6)
+        self.min_cluster.setRange(1, int(1e6))
         self.min_cluster.setValue(10)
         grid.addWidget(self.min_cluster, 0, 1)
         grid.addWidget(QtWidgets.QLabel("Min. samples:"), 1, 0)
         self.min_samples = QtWidgets.QSpinBox()
-        self.min_samples.setRange(1, 1e6)
+        self.min_samples.setRange(1, int(1e6))
         self.min_samples.setValue(10)
         grid.addWidget(self.min_samples, 1, 1)
         grid.addWidget(QtWidgets.QLabel(
@@ -1957,7 +1947,7 @@ class SMLMDialog3D(QtWidgets.QDialog):
         # min no. locs
         grid.addWidget(QtWidgets.QLabel("Min. no. locs:"), 2, 0)
         self.min_locs = QtWidgets.QSpinBox()
-        self.min_locs.setRange(1, 1e6)
+        self.min_locs.setRange(1, int(1e6))
         self.min_locs.setValue(10)
         grid.addWidget(self.min_locs, 2, 1)
         # save cluster centers
@@ -2039,7 +2029,7 @@ class SMLMDialog2D(QtWidgets.QDialog):
         # min no. locs
         grid.addWidget(QtWidgets.QLabel("Min. no. locs:"), 1, 0)
         self.min_locs = QtWidgets.QSpinBox()
-        self.min_locs.setRange(1, 1e6)
+        self.min_locs.setRange(1, int(1e6))
         self.min_locs.setValue(10)
         grid.addWidget(self.min_locs, 1, 1)
         # save cluster centers
@@ -2312,16 +2302,13 @@ class TestClustererDialog(QtWidgets.QDialog):
                 pixelsize,
             )
         elif clusterer_name == "HDBSCAN":
-            if HDBSCAN_IMPORTED:
-                locs = clusterer.hdbscan(
-                    locs,
-                    params["min_cluster_size"],
-                    params["min_samples"],
-                    pixelsize,
-                    params["intercluster_radius"]
-                )
-            else:
-                return None
+            locs = clusterer.hdbscan(
+                locs,
+                params["min_cluster_size"],
+                params["min_samples"],
+                pixelsize,
+                params["intercluster_radius"]
+            )
         elif clusterer_name == "SMLM":
             if params["frame_analysis"]:
                 frame = locs.frame
@@ -2347,9 +2334,14 @@ class TestClustererDialog(QtWidgets.QDialog):
                 ]
 
             locs = clusterer.cluster(locs, params_c, pixelsize)
-          
+        
         if len(locs):
             self.view.group_color = self.window.view.get_group_color(locs)
+
+        # scale z axis if applicable
+        if hasattr(locs, "z"):
+            locs.z /= pixelsize
+        
         return locs
 
     def get_cluster_params(self):
@@ -2401,8 +2393,6 @@ class TestClustererDialog(QtWidgets.QDialog):
         # extract picked locs
         self.channel = self.window.view.get_channel("Test clusterer")
         locs = self.window.view.picked_locs(self.channel)[0]
-        if hasattr(locs, "z"):
-            locs.z /= self.window.display_settings_dlg.pixelsize.value()
         # cluster picked locs
         self.view.locs = self.cluster(locs, params)
         # calculate cluster centers
@@ -2415,14 +2405,10 @@ class TestClustererDialog(QtWidgets.QDialog):
             self.view.viewport = self.view.get_full_fov()
         if self.view.locs is None:
             message = (
-                "No HDBSCAN detected. Please install\n"
-                "the python package HDBSCAN*."
+                "No HDBSCAN detected.\nPlease install the python package"
+                " HDBSCAN (pip install hdbscan)."
             )
-            QtWidgets.QMessageBox.information(
-                self,
-                "No HDBSCAN",
-                message,
-            )
+            QtWidgets.QMessageBox.information(self, "No HDBSCAN", message)
             return
         # render clustered locs
         self.view.update_scene()
@@ -2468,7 +2454,7 @@ class TestDBSCANParams(QtWidgets.QWidget):
         self.min_samples = QtWidgets.QSpinBox()
         self.min_samples.setKeyboardTracking(False)
         self.min_samples.setValue(4)
-        self.min_samples.setRange(1, 1e6)
+        self.min_samples.setRange(1, int(1e6))
         self.min_samples.setSingleStep(1)
         grid.addWidget(self.min_samples, 1, 1)
         grid.setRowStretch(2, 1)
@@ -2487,7 +2473,7 @@ class TestHDBSCANParams(QtWidgets.QWidget):
         self.min_cluster_size = QtWidgets.QSpinBox()
         self.min_cluster_size.setKeyboardTracking(False)
         self.min_cluster_size.setValue(10)
-        self.min_cluster_size.setRange(1, 1e6)
+        self.min_cluster_size.setRange(1, int(1e6))
         self.min_cluster_size.setSingleStep(1)
         grid.addWidget(self.min_cluster_size, 0, 1)
 
@@ -2495,7 +2481,7 @@ class TestHDBSCANParams(QtWidgets.QWidget):
         self.min_samples = QtWidgets.QSpinBox()
         self.min_samples.setKeyboardTracking(False)
         self.min_samples.setValue(10)
-        self.min_samples.setRange(1, 1e6)
+        self.min_samples.setRange(1, int(1e6))
         self.min_samples.setSingleStep(1)
         grid.addWidget(self.min_samples, 1, 1)
 
@@ -2541,7 +2527,7 @@ class TestSMLMParams(QtWidgets.QWidget):
         self.min_locs = QtWidgets.QSpinBox()
         self.min_locs.setKeyboardTracking(False)
         self.min_locs.setValue(10)
-        self.min_locs.setRange(1, 1e6)
+        self.min_locs.setRange(1, int(1e6))
         self.min_locs.setSingleStep(1)
         grid.addWidget(self.min_locs, 2, 1)
 
@@ -2701,57 +2687,61 @@ class TestClustererView(QtWidgets.QLabel):
     def update_scene(self):
         """ Renders localizations. """
 
-        if self.locs is not None:
+        if not len(self.locs):
+            self.setText("No clusters found with the current settings.")
+            return
 
-            if self.viewport is None:
-                self.viewport = self.get_full_fov()
+        if self.viewport is None:
+            self.viewport = self.get_full_fov()
 
-            # split locs according to their group colors
-            locs = self.split_locs()
+        # split locs according to their group colors
+        locs = self.split_locs()
 
-            # render kwargs
-            if self.dialog.one_pixel_blur.isChecked():
-                blur_method = 'smooth'
-            else:
-                blur_method = 'convolve'
-            kwargs = {
-                'oversampling': self.get_optimal_oversampling(),
-                'viewport': self.viewport,
-                'blur_method': blur_method,
-                'min_blur_width': 0,
-                'ang': self.ang,
-            }
+        # render kwargs
+        if self.dialog.one_pixel_blur.isChecked():
+            blur_method = 'smooth'
+        else:
+            blur_method = 'convolve'
+        kwargs = {
+            'oversampling': self.get_optimal_oversampling(),
+            'viewport': self.viewport,
+            'blur_method': blur_method,
+            'min_blur_width': 0,
+            'ang': self.ang,
+        }
 
-            # render images for all channels
-            images = [render.render(_, **kwargs)[1] for _ in locs]
+        # render images for all channels
+        images = [render.render(_, **kwargs)[1] for _ in locs]
 
-            # scale image 
-            images = self.scale_contrast(images)
+        # scale image 
+        images = self.scale_contrast(images)
 
-            # create image to display
-            Y, X = images.shape[1:]
-            bgra = np.zeros((Y, X, 4), dtype=np.float32)
-            colors = get_colors(images.shape[0])
-            for color, image in zip(colors, images): # color each channel
-                bgra[:, :, 0] += color[2] * image
-                bgra[:, :, 1] += color[1] * image
-                bgra[:, :, 2] += color[0] * image
-            bgra = np.minimum(bgra, 1)
-            bgra = self.view.to_8bit(bgra)
-            bgra[:, :, 3].fill(255) # black background
-            qimage = QtGui.QImage(
-                bgra.data, X, Y, QtGui.QImage.Format_RGB32
-            ).scaled(
-                self._size, 
-                self._size, 
-                QtCore.Qt.KeepAspectRatioByExpanding
-            )
-            self.setPixmap(QtGui.QPixmap.fromImage(qimage))
+        # create image to display
+        Y, X = images.shape[1:]
+        bgra = np.zeros((Y, X, 4), dtype=np.float32)
+        colors = get_colors(images.shape[0])
+        for color, image in zip(colors, images): # color each channel
+            bgra[:, :, 0] += color[2] * image
+            bgra[:, :, 1] += color[1] * image
+            bgra[:, :, 2] += color[0] * image
+        bgra = np.minimum(bgra, 1)
+        bgra = self.view.to_8bit(bgra)
+        bgra[:, :, 3].fill(255) # black background
+        qimage = QtGui.QImage(
+            bgra.data, X, Y, QtGui.QImage.Format_RGB32
+        ).scaled(
+            self._size, 
+            self._size, 
+            QtCore.Qt.KeepAspectRatioByExpanding
+        )
+        self.setPixmap(QtGui.QPixmap.fromImage(qimage))
 
     def split_locs(self):
         """
-        Splits self.locs into a list. It has either two, three or 
-        N_GROUP_COLORS elements (each for one group color).
+        Splits self.locs into a list that specifies either separate
+        channels (all localizations, clusters and cluster centers) or
+        it separates clustered localizations by color (based on group,
+        i.e., the cluster id).
         """
 
         if (
@@ -2853,6 +2843,8 @@ class TestClustererView(QtWidgets.QLabel):
             Specifies viewport
         """
 
+        if not len(self.locs):
+            return
         x_min = np.min(self.locs.x) - 1
         x_max = np.max(self.locs.x) + 1
         y_min = np.min(self.locs.y) - 1
@@ -3259,7 +3251,7 @@ class InfoDialog(QtWidgets.QDialog):
             QtWidgets.QLabel("Ignore dark times <="), row, 0
         )
         self.max_dark_time = QtWidgets.QSpinBox()
-        self.max_dark_time.setRange(0, 1e9)
+        self.max_dark_time.setRange(0, int(1e9))
         self.max_dark_time.setValue(1)
         self.picks_grid.addWidget(self.max_dark_time, row, 1, 1, 2)
         row = self.picks_grid.rowCount()
@@ -3279,7 +3271,7 @@ class InfoDialog(QtWidgets.QDialog):
             QtWidgets.QLabel("# Units per pick:"), row, 0
         )
         self.units_per_pick = QtWidgets.QSpinBox()
-        self.units_per_pick.setRange(1, 1e6)
+        self.units_per_pick.setRange(1, int(1e6))
         self.units_per_pick.setValue(1)
         self.picks_grid.addWidget(self.units_per_pick, row, 1, 1, 2)
         calculate_influx_button = QtWidgets.QPushButton("Calibrate influx")
@@ -3532,7 +3524,7 @@ class MaskSettingsDialog(QtWidgets.QDialog):
 
         mask_grid.addWidget(QtWidgets.QLabel("Display pixel size [nm]"), 0, 0)
         self.disp_px_size = QtWidgets.QSpinBox()
-        self.disp_px_size.setRange(0.1, 99999)
+        self.disp_px_size.setRange(10, 99999)
         self.disp_px_size.setValue(300)
         self.disp_px_size.setSingleStep(10)
         self.disp_px_size.setKeyboardTracking(False)
@@ -3898,7 +3890,7 @@ class PickToolCircleSettings(QtWidgets.QWidget):
         self.window = window
         self.grid.addWidget(QtWidgets.QLabel("Diameter (cam. pixel):"), 0, 0)
         self.pick_diameter = QtWidgets.QDoubleSpinBox()
-        self.pick_diameter.setRange(0, 999999)
+        self.pick_diameter.setRange(0.001, 999999)
         self.pick_diameter.setValue(1)
         self.pick_diameter.setSingleStep(0.1)
         self.pick_diameter.setDecimals(3)
@@ -3927,7 +3919,7 @@ class PickToolRectangleSettings(QtWidgets.QWidget):
         self.grid = QtWidgets.QGridLayout(self)
         self.grid.addWidget(QtWidgets.QLabel("Width (cam. pixel):"), 0, 0)
         self.pick_width = QtWidgets.QDoubleSpinBox()
-        self.pick_width.setRange(0, 999999)
+        self.pick_width.setRange(0.001, 999999)
         self.pick_width.setValue(1)
         self.pick_width.setSingleStep(0.1)
         self.pick_width.setDecimals(3)
@@ -4066,7 +4058,7 @@ class RESIDialog(QtWidgets.QDialog):
             self.radius_z.append(r_z)
 
             min_locs = QtWidgets.QSpinBox()
-            min_locs.setRange(1, 1e6)
+            min_locs.setRange(1, int(1e6))
             min_locs.setValue(10)
             min_locs.setSingleStep(1)
             self.min_locs.append(min_locs)
@@ -5056,8 +5048,8 @@ class SlicerDialog(QtWidgets.QDialog):
         self.ax.set_ylabel("Rel. frequency")
         self.ax.set_title(r"$\mathrm{Histogram\ of\ Z:}$")
         self.canvas.draw()
-        self.sl.setMaximum(len(self.bins) - 2)
-        self.sl.setValue(len(self.bins) / 2)
+        self.sl.setMaximum(int(len(self.bins)) - 2)
+        self.sl.setValue(int(len(self.bins) / 2))
 
         # reset cache
         self.slicer_cache = {}
@@ -5071,7 +5063,7 @@ class SlicerDialog(QtWidgets.QDialog):
             self.calculate_histogram()
         else:
             self.calculate_histogram()
-            self.sl.setValue(len(self.bins) / 2)
+            self.sl.setValue(int(len(self.bins) / 2))
             # self.on_slice_position_changed(self.sl.value())
 
     def toggle_slicer(self):
@@ -6045,7 +6037,7 @@ class View(QtWidgets.QLabel):
             locs, 
             radius, 
             min_density,
-            pixelsize,
+            pixelsize=pixelsize,
         )
         dbscan_info = {
             "Generated by": "Picasso DBSCAN",
@@ -6059,7 +6051,7 @@ class View(QtWidgets.QLabel):
         if save_centers:
             status = lib.StatusDialog("Calculating cluster centers", self)
             path = path.replace(".hdf5", "_cluster_centers.hdf5")
-            centers = clusterer.find_cluster_centers(locs, pixelsize)
+            centers = clusterer.find_cluster_centers(locs, pixelsize=pixelsize)
             io.save_locs(path, centers, self.infos[channel] + [dbscan_info])
             status.close()
 
@@ -6067,18 +6059,6 @@ class View(QtWidgets.QLabel):
         """
         Gets channel, parameters and path for HDBSCAN.
         """
-
-        if not HDBSCAN_IMPORTED: # no hdbscan package found
-            message = (
-                "No HDBSCAN detected. Please install\n"
-                "the python package HDBSCAN*."
-            )
-            QtWidgets.QMessageBox.information(
-                self,
-                "No HDBSCAN",
-                message,
-            )
-            return
 
         channel = self.get_channel_all_seq("Cluster")
 
@@ -6152,8 +6132,8 @@ class View(QtWidgets.QLabel):
             locs,
             min_cluster, 
             min_samples, 
-            pixelsize,
-            cluster_eps,
+            pixelsize=pixelsize,
+            cluster_eps=cluster_eps,
         )
         hdbscan_info = {
             "Generated by": "Picasso HDBSCAN",
@@ -6168,7 +6148,7 @@ class View(QtWidgets.QLabel):
         if save_centers:
             status = lib.StatusDialog("Calculating cluster centers", self)
             path = path.replace(".hdf5", "_cluster_centers.hdf5")
-            centers = clusterer.find_cluster_centers(locs, pixelsize)
+            centers = clusterer.find_cluster_centers(locs, pixelsize=pixelsize)
             io.save_locs(path, centers, self.infos[channel] + [hdbscan_info])
             status.close()
 
@@ -6254,7 +6234,9 @@ class View(QtWidgets.QLabel):
                 )
 
                 if len(locs) > 0:
-                    temp_locs = clusterer.cluster(locs, params, pixelsize)
+                    temp_locs = clusterer.cluster(
+                        locs, params, pixelsize=pixelsize
+                    )
 
                     if len(temp_locs) > 0:
                         # make sure each picks produces unique cluster ids
@@ -6279,7 +6261,9 @@ class View(QtWidgets.QLabel):
             else:
                 locs = self.all_locs[channel]
 
-            clustered_locs = clusterer.cluster(locs, params, pixelsize)
+            clustered_locs = clusterer.cluster(
+                locs, params, pixelsize=pixelsize
+            )
             status.close()
 
         # saving
@@ -6429,14 +6413,14 @@ class View(QtWidgets.QLabel):
             alpha = np.arctan((end_y - start_y) / (end_x - start_x))
         dx = width * np.sin(alpha) / 2
         dy = width * np.cos(alpha) / 2
-        x1 = start_x - dx
-        x2 = start_x + dx
-        x4 = end_x - dx
-        x3 = end_x + dx
-        y1 = start_y + dy
-        y2 = start_y - dy
-        y4 = end_y + dy
-        y3 = end_y - dy
+        x1 = int(start_x - dx)
+        x2 = int(start_x + dx)
+        x4 = int(end_x - dx)
+        x3 = int(end_x + dx)
+        y1 = int(start_y + dy)
+        y2 = int(start_y - dy)
+        y4 = int(end_y + dy)
+        y3 = int(end_y - dy)
         return [x1, x2, x3, x4], [y1, y2, y3, y4]
 
     def get_pick_rectangle_polygon(
@@ -6511,6 +6495,7 @@ class View(QtWidgets.QLabel):
             else:
                 d = t_dialog.pick_diameter.value()
                 d *= self.width() / self.viewport_width()
+                d = int(d)
 
                 painter = QtGui.QPainter(image)
                 painter.setPen(QtGui.QColor("yellow"))
@@ -6523,11 +6508,13 @@ class View(QtWidgets.QLabel):
 
                     # convert from camera units to display units
                     cx, cy = self.map_to_view(*pick)
-                    painter.drawEllipse(cx - d / 2, cy - d / 2, d, d)
+                    painter.drawEllipse(int(cx - d / 2), int(cy - d / 2), d, d)
 
                     # annotate picks
                     if t_dialog.pick_annotation.isChecked():
-                        painter.drawText(cx + d / 2, cy + d / 2, str(i))
+                        painter.drawText(
+                            int(cx + d / 2), int(cy + d / 2), str(i)
+                        )
                 painter.end()
 
         # draw rectangular picks
@@ -6645,10 +6632,10 @@ class View(QtWidgets.QLabel):
 
             # draw a cross
             painter.drawPoint(cx, cy)
-            painter.drawLine(cx, cy, cx + d / 2, cy)
-            painter.drawLine(cx, cy, cx, cy + d / 2)
-            painter.drawLine(cx, cy, cx - d / 2, cy)
-            painter.drawLine(cx, cy, cx, cy - d / 2)
+            painter.drawLine(cx, cy, int(cx + d / 2), cy)
+            painter.drawLine(cx, cy, cx, int(cy + d / 2))
+            painter.drawLine(cx, cy, int(cx - d / 2), cy)
+            painter.drawLine(cx, cy, cx, int(cy - d / 2))
 
             # draw a line between points and show distance
             if oldpoint != []:
@@ -6674,7 +6661,9 @@ class View(QtWidgets.QLabel):
                     / 100
                 )
                 painter.drawText(
-                    (cx + ox) / 2 + d, (cy + oy) / 2 + d, str(distance) + " nm"
+                    int((cx + ox) / 2 + d), 
+                    int((cy + oy) / 2 + d), 
+                    str(distance) + " nm",
                 )
             oldpoint = point
         painter.end()
@@ -6763,8 +6752,8 @@ class View(QtWidgets.QLabel):
             painter = QtGui.QPainter(image)
             # initial positions
             x = 12
-            y = 20
-            dy = 20 # space between names
+            y = 26
+            dy = 24 # space between names
             for i in range(n_channels):
                 if self.window.dataset_dialog.checks[i].isChecked():
                     painter.setPen(QtGui.QPen(QtCore.Qt.NoPen))
@@ -6772,7 +6761,7 @@ class View(QtWidgets.QLabel):
                     color = colordisp.palette().color(QtGui.QPalette.Window)
                     painter.setPen(QtGui.QPen(color))
                     font = painter.font()
-                    font.setPixelSize(12)
+                    font.setPixelSize(16)
                     painter.setFont(font)
                     text = self.window.dataset_dialog.checks[i].text()
                     painter.drawText(QtCore.QPoint(x, y), text)
@@ -6797,7 +6786,7 @@ class View(QtWidgets.QLabel):
         if self.window.display_settings_dlg.minimap.isChecked():
             movie_height, movie_width = self.movie_size()
             length_minimap = 100
-            height_minimap = movie_height / movie_width * 100
+            height_minimap = int(movie_height / movie_width * 100)
             # draw in the upper right corner, overview rectangle
             x = self.width() - length_minimap - 20
             y = 20
@@ -6809,10 +6798,10 @@ class View(QtWidgets.QLabel):
             painter.setPen(QtGui.QColor("yellow"))
             if self.window.dataset_dialog.wbackground.isChecked():
                 painter.setPen(QtGui.QColor("red"))
-            length = self.viewport_width() / movie_width * length_minimap
-            height = self.viewport_height() / movie_height * height_minimap
-            x_vp = self.viewport[0][1] / movie_width * length_minimap
-            y_vp = self.viewport[0][0] / movie_height * length_minimap
+            length = int(self.viewport_width() / movie_width * length_minimap)
+            height = int(self.viewport_height() / movie_height * height_minimap)
+            x_vp = int(self.viewport[0][1] / movie_width * length_minimap)
+            y_vp = int(self.viewport[0][0] / movie_height * length_minimap)
             painter.drawRect(x + x_vp, y + y_vp, length + 0, height + 0)
         return image
 
@@ -7299,7 +7288,7 @@ class View(QtWidgets.QLabel):
 
         cx = self.width() * (x - self.viewport[0][1]) / self.viewport_width()
         cy = self.height() * (y - self.viewport[0][0]) / self.viewport_height()
-        return cx, cy
+        return int(cx), int(cy)
 
     def max_movie_height(self):
         """ Returns maximum height of all loaded images. """
@@ -7724,7 +7713,7 @@ class View(QtWidgets.QLabel):
                     fig, (ax1, ax2, ax3) = plt.subplots(
                         3, 1, figsize=(5, 5), constrained_layout=True
                     )
-                    fig.canvas.set_window_title("Trace")
+                    fig.canvas.manager.set_window_title("Trace")
                     pick = self._picks[i]
                     locs = all_picked_locs[i]
                     locs = stack_arrays(locs, asrecarray=True, usemask=False)
@@ -7835,7 +7824,6 @@ class View(QtWidgets.QLabel):
             raise NotImplementedError(
                 "Not implemented for rectangular picks"
             )
-        print("Showing picks...")
         channel = self.get_channel3d("Select Channel")
 
         removelist = [] # picks to be removed
@@ -7855,7 +7843,9 @@ class View(QtWidgets.QLabel):
                     i = 0
                     while i < len(self._picks):
                         fig = plt.figure(figsize=(5, 5))
-                        fig.canvas.set_window_title("Scatterplot of Pick")
+                        fig.canvas.manager.set_window_title(
+                            "Scatterplot of Pick"
+                        )
                         pick = self._picks[i]
 
                         # plot scatter
@@ -7939,7 +7929,9 @@ class View(QtWidgets.QLabel):
                     while i < len(self._picks):
                         pick = self._picks[i]
                         fig = plt.figure(figsize=(5, 5))
-                        fig.canvas.set_window_title("Scatterplot of Pick")
+                        fig.canvas.manager.set_window_title(
+                            "Scatterplot of Pick"
+                        )
                         ax = fig.add_subplot(111)
                         ax.set_title(
                             "Scatterplot of Pick "
@@ -8022,7 +8014,6 @@ class View(QtWidgets.QLabel):
         Uses PlotDialog for displaying the scatter.
         """
 
-        print("Show pick 3D")
         channel = self.get_channel3d("Show Pick 3D")
         pixelsize = self.window.display_settings_dlg.pixelsize.value()
         removelist = []
@@ -8140,7 +8131,6 @@ class View(QtWidgets.QLabel):
     def analyze_cluster(self):
         """ Clusters picked locs using k-means clustering. """
 
-        print("Analyzing clusters...")
         channel = self.get_channel3d("Show Pick 3D")
         removelist = []
         saved_locs = []
@@ -8347,7 +8337,7 @@ class View(QtWidgets.QLabel):
 
                 # plot histogram with n_locs in picks
                 fig = plt.figure()
-                fig.canvas.set_window_title("Localizations in Picks")
+                fig.canvas.manager.set_window_title("Localizations in Picks")
                 ax = fig.add_subplot(111)
                 ax.set_title("Localizations in Picks ")
                 n, bins, patches = ax.hist(
@@ -10061,7 +10051,7 @@ class View(QtWidgets.QLabel):
                 diameter = (
                     self.window.tools_settings_dialog.pick_diameter.value()
                 )
-                diameter = self.width() * diameter / self.viewport_width()
+                diameter = int(self.width() * diameter / self.viewport_width())
                 # remote desktop crashes sometimes for high diameter
                 if diameter < 100:
                     pixmap_size = ceil(diameter)
@@ -10071,7 +10061,7 @@ class View(QtWidgets.QLabel):
                     painter.setPen(QtGui.QColor("white"))
                     if self.window.dataset_dialog.wbackground.isChecked():
                         painter.setPen(QtGui.QColor("black"))
-                    offset = (pixmap_size - diameter) / 2
+                    offset = int((pixmap_size - diameter) / 2)
                     painter.drawEllipse(offset, offset, diameter, diameter)
                     painter.end()
                     cursor = QtGui.QCursor(pixmap)
@@ -10085,7 +10075,7 @@ class View(QtWidgets.QLabel):
         """ Called when evaluating picks statistics in Info Dialog. """
 
         if len(self._picks) == 0:
-            warning = "No picks foundPlease pick first."
+            warning = "No picks found. Please pick first."
             QtWidgets.QMessageBox.information(self, "Warning", warning)
             return
 
@@ -11224,8 +11214,6 @@ class Window(QtWidgets.QMainWindow):
                 self.export_3d_visp()
             elif item == ".csv for ThunderSTORM":
                 self.export_ts()
-            else:
-                print("This should never happen")
 
     def export_ts(self):
         """ Exports locs as .csv for ThunderSTORM. """
@@ -11716,7 +11704,6 @@ class Window(QtWidgets.QMainWindow):
             if input[0] == "flip" and len(input) == 3:
                 # Distinguish flipping in xy and z
                 if "z" in input:
-                    print("xyz")
                     var_1 = input[1]
                     var_2 = input[2]
                     if var_1 == "z":
@@ -11884,7 +11871,6 @@ class Window(QtWidgets.QMainWindow):
         channel = self.view.get_channel_all_seq("Save localizations")
         if channel is not None:
             if channel == len(self.view.locs_paths):
-                print("Save all at once.")
                 suffix, ok = QtWidgets.QInputDialog.getText(
                     self,
                     "Input Dialog",
@@ -12049,6 +12035,7 @@ class Window(QtWidgets.QMainWindow):
         for dialog in self.dialogs:
             dialog.close()
         self.menu_bar.clear() #otherwise the menu bar is doubled
+        self.setWindowTitle("Picasso: Render")
         self.initUI(plugins_loaded=True)
 
     def rot_win(self):